Moved all GUITuitionManager functionality to Main.java & added \n to tostring of studentList when list is empty.
