+++ src/program3/International.java
@@ -1,4 +1,4 @@
-package program2;
+package program3;
 
 /**
  * Class used to represent International Student.
